Rescue from Load Errors
Commit Coveralls-Community/coveralls-ruby@1c349b2 changed LoadError to StandardError to fix some RuboCop
offences.

As per Ruby docs, LoadError is not a subclass of StandardError but it is
a subclass of ScriptError.

When we try to require files inside Coveralls, we are going to rescue
From LoadError as well as StandardError.

Ref:
- https://ruby-doc.org/core-2.6.3/Exception.html