Put cmds in an array
This address issue reported in #105.

`cmds` defaults to a string, which causes the `cmds.each` command to
fail. The only way to manually push is to create a .travis.yml file. To
fix this, wrap the default `cmds` variable in an array, which allows the
`cmds.each` call to successfully call `bundle exec rake` in the case
where a .travis.yml does not exist.